+++ flox/runtime/process/__init__.py
@@ -8,7 +8,7 @@ The former breed of FL Process (namely, synchronous) is the most widely-studied 
 one of the two that (currently) supports hierarchical execution.
 """
 
-from flox.runtime.process.proc_async import AsyncProcess
-from flox.runtime.process.proc_sync import SyncProcess
+from flox.runtime.process.process_async import AsyncProcess
+from flox.runtime.process.process_sync import SyncProcess
 
 __all__ = ["AsyncProcess", "SyncProcess"]
